requirements.txt

Django==1.9.7
six==1.5.2
django-admin-bootstrapped==2.5.7
django-bootstrap3-datetimepicker==2.3
django-celery==3.1.17
django-countries==3.4.1
django-crispy-forms==1.6.0
django-crispy-more==0.0.1
django-cursor-pagination==0.1.1
django-datatables-view==1.13.2
django-debug-toolbar==1.6
django-dynamic-scraper==0.11.0
django-filter==0.15.3
django-filters==0.2.1
django-queryset-csv==1.0.0
django-simple-captcha==0.5.1
djangorestframework==3.5.3
djangorestframework-csv==2.0.0
djangorestframework-filters==0.9.0
djangorestframework-jsonapi==2.1.1
djoser==0.4.3
drf-proxy-pagination==0.1.1
psycopg2==2.6.1


